pyproject.toml
[project]
name = "miniotter"
version = "0.1.0"
description = "GUI-based autonomous agent for macOS"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pyautogui>=0.9.54",
    "pyobjc-framework-ApplicationServices>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "Pillow>=10.0",
    "pywebview>=5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/miniotter"]

[tool.ruff]
line-length = 100
target-version = "py311"
